11/10

+++ app/Http/Controllers/ProjectController.php
@@ -34,41 +34,47 @@ class ProjectController extends Controller
         if ($request->ajax()) {
             return view('projects.projectTablePartial')->with('projects', $projects);
         } 
-    
         return view('projects.index')->with('projects', $projects);
     }
 
-    
-
- 
     public function create()
     {
+        $this->authorize('create', Project::class); 
         return view('projects.create');
     }
-
- 
+    
 
     public function store(CreateProjectRequest $request)
     {
+        $this->authorize('store', Project::class);
         $input = $request->all();
-
         $this->projectRepository->create($input);
-
         return redirect()->route('projects.index')->with('success', 'project added successfully');
-
     }
 
     public function show($id){
         $project = Project::find($id);   
-        return view('projects.view', compact('project'));
+        if($project){
+            return view('projects.view', compact('project'));
+        }else{
+            abort(404);
+        }
     }
+
     public function edit($id){
+        $this->authorize('edit', Project::class);
         $project = Project::find($id);   
-        return view('projects.update', compact('project'));
+        if ($project){
+            return view('projects.update', compact('project'));
+        } else {
+            abort(404);
+        }
+        
     }
 
     public function update(CreateProjectRequest $request, $id)
 {
+    $this->authorize('update', Project::class);
     $project = Project::find($id);
 
     if (!$project) {
@@ -85,10 +91,9 @@ class ProjectController extends Controller
 
 public function destroy($id)
 {
-    // dd($id);
+    $this->authorize('destroy', Project::class);
     Project::find($id)->delete();
     return redirect()->route('projects.index')->with('success', 'Project deleted successfully');
-
 }
 
 }
